database/requirements.txt
psycopg2-binary>=2.9.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0